refactor(steering): extract SteeringHook, remove enable_steering config
Move interruption injection logic from _LoopHook into a standalone
SteeringHook(AgentHook) that plugs into the hooks system. Steering is
now always active — InterruptionChecker is created per-dispatch and
SteeringHook is passed as an extra hook. When no interruptions arrive,
drain_all() returns empty and the hook is a no-op.

Removes enable_steering from AgentDefaults config, CLI wiring, and
AgentLoop.__init__. Removes transform_context/convert_to_llm params
that were unused after the per-tool cancellation drop.

Made-with: Cursor
