+++ nanobot/cli/commands.py
@@ -687,7 +687,6 @@ def gateway(
         unified_session=config.agents.defaults.unified_session,
         disabled_skills=config.agents.defaults.disabled_skills,
         session_ttl_minutes=config.agents.defaults.session_ttl_minutes,
-        enable_steering=config.agents.defaults.enable_steering,
     )
 
     # Set cron callback (needs agent)
@@ -922,7 +921,6 @@ def agent(
         unified_session=config.agents.defaults.unified_session,
         disabled_skills=config.agents.defaults.disabled_skills,
         session_ttl_minutes=config.agents.defaults.session_ttl_minutes,
-        enable_steering=config.agents.defaults.enable_steering,
     )
     restart_notice = consume_restart_notice_from_env()
     if restart_notice and should_show_cli_restart_notice(restart_notice, session_id):
